mealInfoScraping #21 : Separating Meal class from rendering of meal information for modularity and compatibility with other files. Create renderMealInfo function for rendering of arbitrary meals, not just predetermined examples
