+++ frontend/index.html
@@ -6,7 +6,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="profileStyle.css">
-    <script type="module" src="AddMeal.js"></script>
+    <script type="module" src="MealRender.js"></script>
     <link href="mealStyles.css" rel="stylesheet"/>
     <link rel="stylesheet" href="meal-recommender.css">
     <!-- Font Awesome -->
@@ -727,11 +727,7 @@
     
     <div id="mealInfo" class = "view" style="display:none">
         <div class = "momin2">
-            <div class="mealContainer" id = "container">
-                <div class="mealDisplay" id="m1"></div>
-                <div class="mealDisplay" id="m2"></div>
-                <div class="mealDisplay" id="m3"></div>
-                <div class="mealDisplay" id="m4"></div>
+            <div class="mealContainer" id = "mealContainer">
             </div>
         </div>
     </div>
